Move package docs into separate go.doc
